Moved input/output context for multipart/mixed batch to separat files; moved ODataBatchFormat and renamed to ODataMultipartMixedBatchFormat.
